MAIN-18552 - use a less precise number to denote arbitrariness
The number 1997 was fairly arbitrary. It's the largest log line that we've observed being able to print, but there's nothing intrinsic about it that should make it the cutoff. In order to denote that the number is arbitrary, change it to a nice round number like 2048. These lines will still get truncated to ~1997 characters, but are small enought that they won't get dropped, so this does not have any functional consequences.